Major updates, lots of first draft of components
Put in XTAL, did a bunch of research on how to select the capacitors but it comes down to actually picking an oscillator to use, we forgot to order one.
Accidentally ran into the USB micro connector schematic and that was super easy, but we might need to modify for ESD and shielding.
Tried to do the programming pins, but I had trouble finding something on how those work and how we will do it, Clayton can probably handle that another time.
Added linear potentiometer leading into 100Hz lowpass filter, and ESD
Added buttons with ESD